Fix WalletServiceImpl methods

+++ lab_2/src/main/kotlin/it/polito/ecommerce/services/WalletService.kt
@@ -4,12 +4,13 @@ import it.polito.ecommerce.domain.Customer
 import it.polito.ecommerce.dto.CustomerDTO
 import it.polito.ecommerce.dto.TransactionDTO
 import it.polito.ecommerce.dto.WalletDTO
+import java.math.BigDecimal
 import java.sql.Timestamp
 
 interface WalletService {
     fun addWallet(customerID: Int): WalletDTO
     fun getWallet(walletID: Int): WalletDTO
-    fun performTransaction(senderID: Int, receiverID: Int, amount: Double): TransactionDTO
+    fun performTransaction(senderID: Int, receiverID: Int, amount: BigDecimal): TransactionDTO
 //    fun getWalletTransactions(walletID: Int): List<TransactionDTO>
     fun getWalletTransactions(walletID: Int,
                                          from: Long? = null,
